requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
motor==3.1.1
pymongo[srv]==4.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
fuzzywuzzy==0.18.0
python-Levenshtein
python-dotenv